+++ frontend/src/libs/ddah-utils.ts
@@ -56,3 +56,20 @@ export function getRecentActivityDate(ddah: Ddah) {
     }
     return formatDate(new Date(recentActivityDate).toISOString());
 }
+
+/**
+ * Duty descriptions come in the form `<category>:<description>`. This
+ * function breaks a `fullDesc` up into it's two parts
+ *
+ * @param {string} fullDesc
+ * @returns
+ */
+export function splitDutyDescription(fullDesc: string) {
+    const colonPos = fullDesc.indexOf(":");
+    if (colonPos < 0) {
+        return { category: "other", description: fullDesc };
+    }
+    const category = fullDesc.slice(0, colonPos);
+    const description = fullDesc.slice(colonPos + 1);
+    return { category, description };
+}
